refactor: split large errors.ts into focused modules
- Create error-types.ts for type definitions and interfaces (339 lines)
- Create error-factories.ts for error creation functions (199 lines)
- Create error-utilities.ts for guards and aggregation utilities (99 lines)
- Update errors.ts to re-export from focused modules (70 lines)

This addresses the code review feedback about the 599-line errors.ts
file being too large. The new structure organizes error handling by
responsibility (types, factories, utilities) while preserving backward
compatibility through re-exports.
